==> 2_scripts/02_datasheet_summaries_plots.R <==
library(ggplot2)
library(magrittr)


#####################################################################################
output_path <- file.path("3_output", "02_datasheet")

Macroregion_mean_coords_df <- readr::read_csv(file = file.path("1_data", "Macroregion_mean_coords.csv"))
####################################### MAPS ######################################## 

# set manual shapes + colours for 
# region = number
# taxunit = shape 

datasheet_discreteTS_list <- readRDS(file = file.path("1_data","datasheet_discreteTS_list.RDS"))

datasheet_discreteTS_list$sites$TaxUnit_shape <- factor(rep(0, times = nrow(datasheet_discreteTS_list$sites)), levels = c(0:15))

for(current_macroregion in unique(datasheet_discreteTS_list$sites$Macro_region_code)) {
  current_macroregion_subset <- subset(datasheet_discreteTS_list$sites, Macro_region_code == current_macroregion)
  current_unique_TaxUnits <- unique(current_macroregion_subset$TaxUnit)
  
  for (current_unique_TaxUnits_index in 1:length(current_unique_TaxUnits)) {
    a <- current_unique_TaxUnits[current_unique_TaxUnits_index]
    datasheet_discreteTS_list$sites[which(datasheet_discreteTS_list$sites[,"Macro_region_code"] == current_macroregion & datasheet_discreteTS_list$sites[,"TaxUnit"] == a),"TaxUnit_shape"] <- 
      factor(current_unique_TaxUnits_index, levels = c(0:15))
  }
}


# get base map data for extent
world <- rgeos::gBuffer(rworldmap::getMap(resolution = "high"), byid=TRUE, width=0)
clipper_europe <- as(raster::extent(-9.969482, 26.97, 35.74455, 57.5), "SpatialPolygons")
sp::proj4string(clipper_europe) <- sp::CRS(sp::proj4string(world))
world_clip <- raster::intersect(world, clipper_europe)
world_clip_f <- ggplot2::fortify(world_clip)

base_map <- 
  ggplot() +
  geom_polygon(data = world_clip_f,
               aes(x = long, y = lat, group = group),
               fill = NA, colour = "grey") +
  coord_map() +
  theme_classic() +  
  xlab("Longitude") +
  ylab("Latitude") +
  scale_y_continuous(expand = c(0,0)) + 
  scale_x_continuous(expand = c(0,0)) +
  theme(legend.position = "right",
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 12),
        legend.text = element_text(size = 14),
        legend.title = element_text(size = 16)) + 
  theme(text = element_text(family = "-monotype-arial-%s-%s-*-*-%d-*-*-*-*-*-*-*"))



############################################## COUNTS ############################################## 

# taxunits per timeslice from sites sheet
TaxUnits_per_TS_count_df_list <- list()
for(current_timeslice in unique(datasheet_discreteTS_list$sites$Timeslice_discrete)) {
  
  current_timeslice_subset <- subset(datasheet_discreteTS_list$sites, 
                                     Timeslice_discrete == current_timeslice)
  
  TaxUnits_per_TS_count_df_list[[current_timeslice]] <- 
    data.frame(Timeslice = current_timeslice,
               N_unique_TaxUnits = length(unique(current_timeslice_subset$TaxUnit)), 
               TaxUnits = paste0(unique(current_timeslice_subset$TaxUnit), collapse = ","))
}
TaxUnits_per_TS_count_df <-  do.call(rbind.data.frame, TaxUnits_per_TS_count_df_list)


TaxUnits_per_TS_count_df$Timeslice <- gsub("1", "I", TaxUnits_per_TS_count_df$Timeslice )
TaxUnits_per_TS_count_df$Timeslice  <- gsub("2", "II", TaxUnits_per_TS_count_df$Timeslice )
TaxUnits_per_TS_count_df$Timeslice  <- gsub("3", "III", TaxUnits_per_TS_count_df$Timeslice )
TaxUnits_per_TS_count_df$Timeslice  <- gsub("4", "IV", TaxUnits_per_TS_count_df$Timeslice )

TaxUnits_per_TS_count_df$Timeslice <- factor(TaxUnits_per_TS_count_df$Timeslice, 
                                             levels = c("I", "II", "III", "IV"))

TaxUnits_per_TS_count_df[,c(1,2)]

readr::write_csv(TaxUnits_per_TS_count_df,
                 file = file.path(output_path, "TaxUnits_per_TS.csv"))

TaxUnits_per_TS_plot <- 
  ggplot(data = TaxUnits_per_TS_count_df) +
  geom_col(aes(x = Timeslice, 
               y = N_unique_TaxUnits,
               fill = Timeslice)) +
  ylab("Number of unique NACs") +
  theme_bw() +
  xlab("Time-slice") +
  scale_fill_manual(values = c("I" = "#86b1c0",
                               "II" = "#8694c0",
                               "III" = "#9586c0",
                               "IV" = "#b286c0")) +
  theme(legend.position = "none") + 
  theme(text = element_text(family = "-monotype-arial-%s-%s-*-*-%d-*-*-*-*-*-*-*"))

TaxUnits_per_TS_plot

ggsave(plot = TaxUnits_per_TS_plot,
       filename = file.path(output_path , "SI_Figure_1_TaxUnits_per_TS.png"), 
       width = 10, height = 10, units = "cm")



# taxunits per macroregion
TaxUnits_per_Region_count_df_list <- list()
for(current_macroregion in unique(datasheet_discreteTS_list$sites$Macro_region_code)) {
  
  current_macroregion_subset <- subset(datasheet_discreteTS_list$sites, 
                                       Macro_region_code == current_macroregion)
  
  TaxUnits_per_Region_count_df_list[[current_macroregion]] <- 
    data.frame(Macroregion = current_macroregion,
               N_unique_TaxUnits = length(unique(current_macroregion_subset$TaxUnit)), 
               TaxUnits = paste0(unique(current_macroregion_subset$TaxUnit), collapse = ","))
}
TaxUnits_per_Region_count_df <-  do.call(rbind.data.frame, TaxUnits_per_Region_count_df_list)


TaxUnits_per_Region_count_df$Macroregion <- factor(TaxUnits_per_Region_count_df$Macroregion,
                                                   levels = Macroregion_mean_coords_df$Macro_region_code[order(Macroregion_mean_coords_df$Macro_region_code_mean_Lat, 
                                                                                                               decreasing = T)])

TaxUnits_per_Region_count_df <- na.omit(TaxUnits_per_Region_count_df)

TaxUnits_per_Region_plot <- 
  ggplot(data = TaxUnits_per_Region_count_df) +
  geom_col(aes(x = Macroregion, 
               y = N_unique_TaxUnits
               )) +
  theme_bw() +
  xlab("Region") + 
  ylab("Number of unique NACs") +
  theme(legend.position = "none") + 
  theme(text = element_text(family = "-monotype-arial-%s-%s-*-*-%d-*-*-*-*-*-*-*"))

TaxUnits_per_Region_plot

# ggsave(plot = TaxUnits_per_Region_plot,
#        filename = file.path(output_path, "TaxUnits_per_Region.png"), 
#        width = 20, height = 10, units = "cm")


# taxunits per macroregion per timeslice
TaxUnits_per_Region_per_TS_count_df_list <- list()

for(current_timeslice in unique(datasheet_discreteTS_list$sites$Timeslice_discrete)) {
  
  current_timeslice_subset <- subset(datasheet_discreteTS_list$sites, 
                                     Timeslice_discrete == current_timeslice)
  
  TaxUnits_per_Region_count_df_list <- list()
  for(current_macroregion in unique(current_timeslice_subset$Macro_region_code)) {
    
    current_macroregion_subset <- subset(current_timeslice_subset, 
                                         Macro_region_code == current_macroregion)
    
    TaxUnits_per_Region_count_df_list[[current_macroregion]] <- 
      data.frame(Timeslice = current_timeslice,
                 Macroregion = current_macroregion,
                 N_unique_TaxUnits = length(unique(current_macroregion_subset$TaxUnit)), 
                 TaxUnits = paste0(unique(current_macroregion_subset$TaxUnit), collapse = ","),
                 TaxUnits_single = unique(current_macroregion_subset$TaxUnit))
  }
  TaxUnits_per_Region_per_TS_count_df_list[[current_timeslice]] <-  do.call(rbind.data.frame, TaxUnits_per_Region_count_df_list)
  
  
}
TaxUnits_per_Region_per_TS_count_df <-  do.call(rbind.data.frame, TaxUnits_per_Region_per_TS_count_df_list)

TaxUnits_per_Region_per_TS_count_df$Macroregion <- factor(TaxUnits_per_Region_per_TS_count_df$Macroregion,
                                                          levels = Macroregion_mean_coords_df$Macro_region_code[order(Macroregion_mean_coords_df$Macro_region_code_mean_Lat, 
                                                                                                                      decreasing = T)])
TaxUnits_per_Region_per_TS_count_df <- na.omit(TaxUnits_per_Region_per_TS_count_df)

TaxUnits_per_Region_per_TS_count_df$Timeslice <- gsub("1", "I", TaxUnits_per_Region_per_TS_count_df$Timeslice )
TaxUnits_per_Region_per_TS_count_df$Timeslice  <- gsub("2", "II", TaxUnits_per_Region_per_TS_count_df$Timeslice )
TaxUnits_per_Region_per_TS_count_df$Timeslice  <- gsub("3", "III", TaxUnits_per_Region_per_TS_count_df$Timeslice )
TaxUnits_per_Region_per_TS_count_df$Timeslice  <- gsub("4", "IV", TaxUnits_per_Region_per_TS_count_df$Timeslice )

TaxUnits_per_Region_per_TS_count_df$Timeslice <- factor(TaxUnits_per_Region_per_TS_count_df$Timeslice, 
                                             levels = c("I", "II", "III", "IV"))

TaxUnits_per_Region_per_TS_plot <- 
  ggplot(data = unique(dplyr::select(TaxUnits_per_Region_per_TS_count_df, -TaxUnits_single))) +
  geom_col(aes(x = Macroregion, 
               y = N_unique_TaxUnits,
               fill = Timeslice
               )) +
  facet_wrap(~Timeslice) +
  theme_bw() +
  ylab("Number of unique NACs") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("I" = "#86b1c0",
                               "II" = "#8694c0",
                               "III" = "#9586c0",
                               "IV" = "#b286c0")) + 
  theme(text = element_text(family = "-monotype-arial-%s-%s-*-*-%d-*-*-*-*-*-*-*"))

TaxUnits_per_Region_per_TS_plot

# ggsave(plot = TaxUnits_per_Region_per_TS_plot,
#        filename = file.path(output_path, "Fig_1C_TaxUnits_per_Region_per_TS.png"), 
#        width = 30, height = 15, units = "cm")
# svg(filename=file.path(output_path, "Fig_1C_TaxUnits_per_Region_per_TS.svg"), width = 30*0.39, height = 15*0.39);TaxUnits_per_Region_per_TS_plot;dev.off()


N_unique_taxUnits_per_TS <-
  TaxUnits_per_Region_per_TS_count_df %>% 
    dplyr::select(., -c(TaxUnits, N_unique_TaxUnits)) %>% 
    unique() %>% 
    dplyr::group_by(Timeslice) %>% 
    dplyr::mutate(N_TaxUnits_per_Timeslice = length(unique(TaxUnits_single))) %>% 
    dplyr::select(-c(TaxUnits_single, Macroregion)) %>% 
    unique()

unique_taxUnits_per_Region <-
  TaxUnits_per_Region_per_TS_count_df %>% 
    dplyr::select(., -c(TaxUnits, N_unique_TaxUnits)) %>% 
    unique() %>% 
    dplyr::group_by(Macroregion) %>% 
    dplyr::mutate(N_TaxUnits_per_Region = length(unique(TaxUnits_single))) 

N_unique_taxUnits_per_Region <- 
  unique_taxUnits_per_Region %>% 
  dplyr::select(-c(Timeslice, TaxUnits_single)) %>% 
  unique()


TaxUnits_per_Region_per_TS_count_df_noSingle <- 
  TaxUnits_per_Region_per_TS_count_df %>% 
  dplyr::select(-c(TaxUnits_single, N_unique_TaxUnits)) %>% 
  unique()

table_1_regions <- 
  reshape2::dcast(unique(TaxUnits_per_Region_per_TS_count_df_noSingle), Macroregion ~ Timeslice) %>% 
  dplyr::left_join(.,
                   N_unique_taxUnits_per_Region
                   )

n_NACs_per_TS <- N_unique_taxUnits_per_TS$N_TaxUnits_per_Timeslice

table_1_regions[nrow(table_1_regions)+1,] <- c(NA, 
                                               n_NACs_per_TS,
                                               NA)

readr::write_csv(table_1_regions,
                 file = file.path(output_path, "table_1_unique_N_TaxUnits_perRegion_perTS.csv"))


##########################################
# quality scores
##########################################

##################### map with site types, quality scores
map_all_sites_SiteType_Quali <- 
  base_map + 
  labs(fill = "Quality Score",
       shape = "Site Type") +
  geom_point(data = datasheet_discreteTS_list$sites,
             aes(x = Long, y = Lat,
                 fill = quality_score,
                 shape = Site_type),
             alpha = 0.9,
             size = 3) +
  scale_fill_gradient2(low = "red", mid = "yellow", high = "darkgreen",
                       midpoint = 3)+
  scale_shape_manual(values=c(21, 22, 24)) + 
  theme(text = element_text(family = "-monotype-arial-%s-%s-*-*-%d-*-*-*-*-*-*-*"))

map_all_sites_SiteType_Quali

ggsave(plot = map_all_sites_SiteType_Quali,
       filename = file.path(output_path, "SI_Fig_2A_map_all_sites_siteType_quali.png"), 
       width = 30, height = 24, units = "cm")
svg(filename=file.path(output_path, "SI_Fig_2A_map_all_sites_siteType_quali.svg"), width = 30*0.39, height = 24*0.39);map_all_sites_SiteType_Quali;dev.off()

#################### quality score by Region by Timeslice
quality_score_by_ts_by_region_list <- list()
for (ts_index in unique(datasheet_discreteTS_list[["sites"]]$Timeslice_discrete)) {
  subset_by_ts <- subset(datasheet_discreteTS_list[["sites"]], Timeslice_discrete == ts_index)
  subset_by_ts <- unique(subset_by_ts)
  
  quality_score_by_region_list <- list()
  for (current_macroregion in unique(subset_by_ts$Macro_region_code)) {
    subset_by_region <- subset(subset_by_ts, Macro_region_code == current_macroregion)
    subset_by_region <- unique(subset_by_region)
    
    quality_score_by_region_list[[current_macroregion]] <-
      data.frame(Macro_region_code = factor(current_macroregion, 
                                            levels =  Macroregion_mean_coords_df$Macro_region_code[order(Macroregion_mean_coords_df$Macro_region_code_mean_Lat, decreasing = T)]),
                 Macro_region_code_n = factor(paste0(current_macroregion, 
                                                     "\n(n=", 
                                                     nrow(subset_by_region), 
                                                     ")")),
                 TS_discrete = ts_index,
                 quality_median_score = round(median(subset_by_region$quality_score), 
                                   digits = 2)) # quality score
    
  }
  temp_df <- do.call(rbind.data.frame, quality_score_by_region_list)
  quality_score_by_ts_by_region_list[[ts_index]] <- temp_df
}
quality_score_by_ts_by_region_df <- do.call(rbind.data.frame, quality_score_by_ts_by_region_list)

quality_score_by_ts_by_region_df <- dplyr::left_join(quality_score_by_ts_by_region_df,
                                                     datasheet_discreteTS_list$sites[,c("Macro_region_code", "Macro_region_number", "quality_score", "Timeslice_discrete")],
                                                     by = c("Macro_region_code", "TS_discrete" = "Timeslice_discrete"))

quality_score_by_ts_by_region_df$TS_discrete <- gsub(1, "I", quality_score_by_ts_by_region_df$TS_discrete)
quality_score_by_ts_by_region_df$TS_discrete <- gsub(2, "II", quality_score_by_ts_by_region_df$TS_discrete)
quality_score_by_ts_by_region_df$TS_discrete <- gsub(3, "III", quality_score_by_ts_by_region_df$TS_discrete)
quality_score_by_ts_by_region_df$TS_discrete <- gsub(4, "IV", quality_score_by_ts_by_region_df$TS_discrete)

quality_score_by_ts_by_region_plot <- 
ggplot(data = quality_score_by_ts_by_region_df) + 
  ggpointgrid::geom_pointrect(aes(x = Macro_region_code,
                                  y = quality_score,
                                  fill = quality_score),
                              shape = 21,
                              # size = 1,
                              scale_x = 0.2,
                              scale_y = 0.1) +
  geom_point(aes(x = Macro_region_code,
                 y = quality_median_score),
             color = "black",
             shape = 21,
             size = 8) +
  theme_bw() +
  ylab("Quality score") +
  xlab("Macroregion") +
  facet_wrap(~TS_discrete) +
  scale_fill_gradient2(low = "red", mid = "yellow", high = "darkgreen",
                       midpoint = 3) +
  theme(legend.position = "none",
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank()) +
  scale_y_continuous(breaks=c(0:6), limits = c(0, 6)) + 
  theme(text = element_text(family = "-monotype-arial-%s-%s-*-*-%d-*-*-*-*-*-*-*"))

quality_score_by_ts_by_region_plot

quality_score_by_ts_by_region_df_medians <- unique(quality_score_by_ts_by_region_df %>% 
                                                     dplyr::select(-c("Macro_region_number", "quality_score")))
quality_score_by_ts_by_region_df_medians
readr::write_csv(quality_score_by_ts_by_region_df_medians,
                 file = file.path(output_path, "quality_score_by_ts_by_region_df_medians.csv"))

ggsave(plot = quality_score_by_ts_by_region_plot,
       filename = file.path(output_path, "SI_Fig_2B_quality_score_by_ts_by_region.png"), 
       width = 30, height = 20, units = "cm")
svg(filename=file.path(output_path, "SI_Fig_2B_quality_score_by_ts_by_region.svg"), width = 30*0.39, height = 20*0.39);quality_score_by_ts_by_region_plot;dev.off()
